cleaner/requirements.txt
flask
flask-restful
nltk
pandas
numpy
textblob
kafka-python
